[Fix] 카카오 로그인 이메일 필수 동의로 로직 수정 및 로그 정리
