+++ src/main/java/wowmarket/wow_server/login/service/KakaoAPIService.java
@@ -1,6 +1,5 @@
 package wowmarket.wow_server.login.service;
 
-
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
@@ -15,7 +14,6 @@ import wowmarket.wow_server.repository.UserRepository;
 import java.io.*;
 import java.net.HttpURLConnection;
 import java.net.URL;
-import java.net.URLConnection;
 import java.util.HashMap;
 import java.util.Optional;
 
@@ -25,7 +23,7 @@ import java.util.Optional;
  */
 @Service
 @RequiredArgsConstructor
-public class KakaoAPI {
+public class KakaoAPIService {
     private final UserRepository userRepository;
 
     public String getAccessToken(String authorize_code) {
@@ -103,35 +101,19 @@ public class KakaoAPI {
 
             JsonParser parser = new JsonParser();
             JsonElement element = parser.parse(result);
-            System.out.println("[getUserInfo] element는 뭘까.. element = " + element);
 
             JsonObject properties = element.getAsJsonObject().get("properties").getAsJsonObject();
             JsonObject kakao_account = element.getAsJsonObject().get("kakao_account").getAsJsonObject();
 
-            System.out.println("[getUserInfo] JsonObject kakao_account로 넘어오는 값은 뭘까 kakao_account = " + kakao_account);
-            System.out.println("[getUserInfo] 그러면 JsonObject properties는 뭘까 properties = " + properties);
-            System.out.println("[getUserInfo] 이메일 동의 필요한지 확인 kakao_account.getAsJsonObject().get(\"email_needs_agreement\").getAsString() : " + kakao_account.getAsJsonObject().get("email_needs_agreement").getAsString());
-
             String nickname = properties.getAsJsonObject().get("nickname").getAsString();
-            //email_needs_agreement로 이메일 동의 했는지 확인
-            //동의 안 했으면 빈 해쉬값이 넘어가서 controller에서 연결 끊기 api가 실행되고 다시 로그인 하도록 설정
-            //email_needs_agreement 가 true 이면 이메일 정보 제공에 동의를 안 해서 동의가 필요하다는 뜻
-            if (kakao_account.getAsJsonObject().get("email_needs_agreement").getAsString() == "true") {
-                System.out.println("[getUserInfo] 이메일 제공 미동의로 [getUserInfo] 종료");
-                return userInfo;
-            }
             String email = kakao_account.getAsJsonObject().get("email").getAsString();
 
             System.out.println("[getUserInfo] nickname = " + nickname);
             System.out.println("[getUserInfo] email = " + email);
-//
-//            System.out.println("\n\n **** userRepository.findByEmail(email).isEmpty() : " + userRepository.findByEmail(email).isEmpty());
-//            System.out.println("userRepository.findByEmail(email) : " + userRepository.findByEmail(email) + "\n\n");
-            Optional<User> userFindbyEmail = userRepository.findByEmail(email);
+
             //findByEmail로 값이 없으면 DB에 저장 userRepository.findByEmail(email).isEmpty()
-            if (userFindbyEmail.isEmpty()) { // 쿼리 너무 많이 나가서 -> 추후에 userRepository.findByEmail(email).isEmpty() 로 교체
-                System.out.println("\n[getUserInfo] userId와 email 값 잘 넘어왔고 DB에 해당 email 없어서 DB에 저장하는 로직 실행");
-                System.out.println("[getUserInfo] DB 저장 전 Optional<User> userFindbyEmail = userRepository.findByEmail(email) : " + userFindbyEmail);
+            if (userRepository.findByEmail(email).isEmpty()) {
+                System.out.println("\n[getUserInfo] nickname과 email 값 잘 넘어왔고 DB에 해당 email 없어서 DB에 저장하는 로직 실행");
                 KakaoDto kakaoDto = new KakaoDto(email, nickname, Login_Method.KAKAO);
                 User User = userRepository.save(kakaoDto.toEntity());
                 System.out.println("[getUserInfo] DB 저장 후 User 테이블 userId 확인 userRepository.findByEmail(email).get().getId() : " + userRepository.findByEmail(email).get().getId());
@@ -142,6 +124,7 @@ public class KakaoAPI {
             userInfo.put("nickname", nickname);
             userInfo.put("email", email);
 
+            br.close();
         } catch (IOException e) {
             e.printStackTrace();
         }
@@ -167,15 +150,15 @@ public class KakaoAPI {
             while ((line = br.readLine()) != null) {
                 result += line;
             }
-            System.out.println("[unLink] response body : " + result + " 연결 끊기에 성공한 사용자의 회원번호 이 회원번호는 카카오의 회원번호일까? 근데 매번 바뀌는디");
-            //해당 id는 연결 끊기에 성공한 사용자의 회원번호
+            System.out.println("[unLink] response body : " + result + " 연결 끊기에 성공한 사용자의 회원번호, 이 회원번호는 카카오의 회원번호일까? 근데 매번 바뀌는 거 같은디");
 
+            br.close();
         } catch (IOException e) {
             e.printStackTrace();
         }
     }
 
-    public void kakaoLogout(String acccess_token) {
+    public void logout(String acccess_token) {
         String reqURL = "https://kapi.kakao.com/v1/user/logout";
 
         try {
@@ -185,7 +168,7 @@ public class KakaoAPI {
             conn.setRequestProperty("Authorization", "Bearer " + acccess_token);
 
             int responseCode = conn.getResponseCode();
-            System.out.println("로그아웃 상태코드 반환 : " + responseCode);
+            System.out.println("\n[logout] 상태코드 반환 : " + responseCode);
 
             BufferedReader br = new BufferedReader(new InputStreamReader(conn.getInputStream()));
             String line = "";
@@ -193,8 +176,9 @@ public class KakaoAPI {
             while ((line = br.readLine()) != null) {
                 result += line;
             }
-            System.out.println("로그아웃 response body : " + result);
+            System.out.println("[logout] response body : " + result);
 
+            br.close();
         } catch (IOException e) {
             e.printStackTrace();
         }
