+++ src/main/resources/templates/index.html
@@ -1,37 +1,22 @@
-<!--<html xmlns:th="http://www.thymeleaf.org">-->
-<!--<body>-->
-<!--&lt;!&ndash;<p th:text="'hello ' + ${name}">hello! empty</p>&ndash;&gt;-->
-<!--<span th:if="${userId == null}">-->
-<!--    <a href="https://kauth.kakao.com/oauth/authorize?client_id=394cbd2e5e0ad400adbc202784ad624b&redirect_uri=http://localhost:8080/wowmarket/users/login/kakao&response_type=code">-->
-<!--        <img src="img/kakao_login_medium_narrow.png">-->
-<!--    </a>-->
-<!--</span>-->
-<!--<span th:if="${userId != null}">-->
-<!--    <h1>로그인 성공</h1>-->
-<!--    <input type="button" value="로그아웃" onclick="location.href='/logout'">-->
-<!--</span>-->
-<!--</body>-->
-<!--</html>-->
-
 <html xmlns:th="http://www.thymeleaf.org">
 <body>
 <div id="loginContent">
-        <span th:if="${userId == null}">
-            <a href="https://kauth.kakao.com/oauth/authorize?client_id=394cbd2e5e0ad400adbc202784ad624b&redirect_uri=http://localhost:8080/wowmarket/users/login/kakao&response_type=code">
-                <img src="img/kakao_login_medium_narrow.png">
-            </a>
-            <a href="http://localhost:8080/wowmarket/users/kakao/unlink">
-                <button value="연결끊기">연결끊기</button>
-            </a>
-        </span>
+    <span th:if="${session.user_email == null}">
+        <a href="https://kauth.kakao.com/oauth/authorize?client_id=394cbd2e5e0ad400adbc202784ad624b&redirect_uri=http://localhost:8080/wowmarket/users/login/kakao&response_type=code">
+            <img src="img/kakao_login_medium_narrow.png">
+        </a>
+    </span>
+    <span th:if="${session.user_email != null}">
+        <b>[로그인 성공]</b>
+        <br><br>
+        <a href="http://localhost:8080/wowmarket/users/kakao/unlink">
+            <button value="연결끊기">연결끊기</button>
+        </a>
+        <br><br>
+        <a href="http://localhost:8080/wowmarket/users/logout/kakao">
+            <button value="로그아웃">로그아웃</button>
+        </a>
+    </span>
 </div>
-
-<script th:inline="javascript">
-        /* 타임리프 변수를 JavaScript 변수로 사용 */
-        var userId = /*[[${userId}]]*/ null;
-</script>
-
-<!-- 외부 JavaScript 파일 로드 -->
-<script src="js/main.js"></script>
 </body>
 </html>
